+++ minion/scheduler/worker.go
@@ -182,7 +182,9 @@ func doContainers(dk docker.Client, ifaces []interface{},
 func dockerRun(dk docker.Client, iface interface{}) {
 	dbc := iface.(evaluatedContainer)
 	log.WithField("container", dbc).Info("Start container")
+
 	_, err := dk.Run(docker.RunOptions{
+		Hostname:          dbc.Hostname + ".q",
 		Image:             dbc.Image,
 		Args:              dbc.Command,
 		Env:               dbc.resolvedEnv,
@@ -220,7 +222,8 @@ func syncJoinScore(left, right interface{}) int {
 	dkc := right.(docker.Container)
 
 	expFilesHash := filesHash(dbc.resolvedFilepathToContent)
-	if dbc.IP != dkc.IP || expFilesHash != dkc.Labels[filesKey] {
+	if dbc.Hostname+".q" != dkc.Hostname || dbc.IP != dkc.IP ||
+		expFilesHash != dkc.Labels[filesKey] {
 		return -1
 	}